+++ Tests/FetchRequestTests.swift
@@ -3,9 +3,9 @@ import GRDB
 import RxSwift
 @testable import RxGRDB // @testable to get PrimaryKeyDiff initializer
 
-class TypedRequestTests : XCTestCase { }
+class FetchRequestTests : XCTestCase { }
 
-extension TypedRequestTests {
+extension FetchRequestTests {
     func setUpDatabase(in writer: DatabaseWriter) throws {
         try writer.write { db in
             try db.create(table: "players") { t in
@@ -40,7 +40,7 @@ extension TypedRequestTests {
 
 // MARK: - RowConvertible
 
-extension TypedRequestTests {
+extension FetchRequestTests {
     func testRxFetchAllRecords() throws {
         try Test(testRxFetchAllRecords)
             .run { try DatabaseQueue(path: $0) }
@@ -50,7 +50,6 @@ extension TypedRequestTests {
     func testRxFetchAllRecords(writer: DatabaseWriter, disposeBag: DisposeBag) throws {
         let request = Player.order(Column("name"))
         let expectedNames = [
-            ["Arthur", "Barbara"],
             ["Arthur", "Barbara"],
             ["Arthur", "Barbie"],
             [],
@@ -75,41 +74,7 @@ extension TypedRequestTests {
     }
 }
 
-extension TypedRequestTests {
-    func testRxFetchAllRecordsDistinctUntilChanged() throws {
-        try Test(testRxFetchAllRecordsDistinctUntilChanged)
-            .run { try DatabaseQueue(path: $0) }
-            .run { try DatabasePool(path: $0) }
-    }
-    
-    func testRxFetchAllRecordsDistinctUntilChanged(writer: DatabaseWriter, disposeBag: DisposeBag) throws {
-        let request = Player.order(Column("name"))
-        let expectedNames = [
-            ["Arthur", "Barbara"],
-            ["Arthur", "Barbie"],
-            [],
-            ["Craig", "David", "Elena"],
-            ]
-        
-        try setUpDatabase(in: writer)
-        let recorder = EventRecorder<[Player]>(expectedEventCount: expectedNames.count)
-        request.rx.fetchAll(in: writer, distinctUntilChanged: true)
-            .subscribe { event in
-                // events are expected on the main thread by default
-                assertMainQueue()
-                recorder.on(event)
-            }
-            .disposed(by: disposeBag)
-        try modifyDatabase(in: writer)
-        wait(for: recorder, timeout: 1)
-        
-        for (event, names) in zip(recorder.recordedEvents, expectedNames) {
-            XCTAssertEqual(event.element!.map { $0.name }, names)
-        }
-    }
-}
-
-extension TypedRequestTests {
+extension FetchRequestTests {
     func testRxFetchAllRecordsIdentifiedByIds() throws {
         try Test(testRxFetchAllRecordsIdentifiedByIds)
             .run { try DatabaseQueue(path: $0) }
@@ -143,7 +108,7 @@ extension TypedRequestTests {
     }
 }
 
-extension TypedRequestTests {
+extension FetchRequestTests {
     func testRxFetchOneRecord() throws {
         try Test(testRxFetchOneRecord)
             .run { try DatabaseQueue(path: $0) }
@@ -153,8 +118,6 @@ extension TypedRequestTests {
     func testRxFetchOneRecord(writer: DatabaseWriter, disposeBag: DisposeBag) throws {
         let request = Player.order(Column("name"))
         let expectedNames = [
-            "Arthur",
-            "Arthur",
             "Arthur",
             nil,
             "Craig",
@@ -178,40 +141,7 @@ extension TypedRequestTests {
     }
 }
 
-extension TypedRequestTests {
-    func testRxFetchOneRecordDistinctUntilChanged() throws {
-        try Test(testRxFetchOneRecordDistinctUntilChanged)
-            .run { try DatabaseQueue(path: $0) }
-            .run { try DatabasePool(path: $0) }
-    }
-    
-    func testRxFetchOneRecordDistinctUntilChanged(writer: DatabaseWriter, disposeBag: DisposeBag) throws {
-        let request = Player.order(Column("name"))
-        let expectedNames = [
-            "Arthur",
-            nil,
-            "Craig",
-            ]
-        
-        try setUpDatabase(in: writer)
-        let recorder = EventRecorder<Player?>(expectedEventCount: expectedNames.count)
-        request.rx.fetchOne(in: writer, distinctUntilChanged: true)
-            .subscribe { event in
-                // events are expected on the main thread by default
-                assertMainQueue()
-                recorder.on(event)
-            }
-            .disposed(by: disposeBag)
-        try modifyDatabase(in: writer)
-        wait(for: recorder, timeout: 1)
-        
-        for (event, name) in zip(recorder.recordedEvents, expectedNames) {
-            XCTAssertEqual(event.element!?.name, name)
-        }
-    }
-}
-
-extension TypedRequestTests {
+extension FetchRequestTests {
     func testRxFetchOneRecordIdentifiedById() throws {
         try Test(testRxFetchOneRecordIdentifiedById)
             .run { try DatabaseQueue(path: $0) }
@@ -247,7 +177,7 @@ extension TypedRequestTests {
 
 // MARK: - Row
 
-extension TypedRequestTests {
+extension FetchRequestTests {
     func testRxFetchAllRows() throws {
         try Test(testRxFetchAllRows)
             .run { try DatabaseQueue(path: $0) }
@@ -257,7 +187,6 @@ extension TypedRequestTests {
     func testRxFetchAllRows(writer: DatabaseWriter, disposeBag: DisposeBag) throws {
         let request = SQLRequest<Row>("SELECT * FROM players ORDER BY name")
         let expectedNames = [
-            ["Arthur", "Barbara"],
             ["Arthur", "Barbara"],
             ["Arthur", "Barbie"],
             [],
@@ -282,41 +211,7 @@ extension TypedRequestTests {
     }
 }
 
-extension TypedRequestTests {
-    func testRxFetchAllRowsDistinctUntilChanged() throws {
-        try Test(testRxFetchAllRowsDistinctUntilChanged)
-            .run { try DatabaseQueue(path: $0) }
-            .run { try DatabasePool(path: $0) }
-    }
-    
-    func testRxFetchAllRowsDistinctUntilChanged(writer: DatabaseWriter, disposeBag: DisposeBag) throws {
-        let request = SQLRequest<Row>("SELECT * FROM players ORDER BY name")
-        let expectedNames = [
-            ["Arthur", "Barbara"],
-            ["Arthur", "Barbie"],
-            [],
-            ["Craig", "David", "Elena"],
-            ]
-        
-        try setUpDatabase(in: writer)
-        let recorder = EventRecorder<[Row]>(expectedEventCount: expectedNames.count)
-        request.rx.fetchAll(in: writer, distinctUntilChanged: true)
-            .subscribe { event in
-                // events are expected on the main thread by default
-                assertMainQueue()
-                recorder.on(event)
-            }
-            .disposed(by: disposeBag)
-        try modifyDatabase(in: writer)
-        wait(for: recorder, timeout: 1)
-        
-        for (event, names) in zip(recorder.recordedEvents, expectedNames) {
-            XCTAssertEqual(event.element!.map { $0["name"] as String }, names)
-        }
-    }
-}
-
-extension TypedRequestTests {
+extension FetchRequestTests {
     func testRxFetchAllRowsIdentifiedByIds() throws {
         try Test(testRxFetchAllRowsIdentifiedByIds)
             .run { try DatabaseQueue(path: $0) }
@@ -350,7 +245,7 @@ extension TypedRequestTests {
     }
 }
 
-extension TypedRequestTests {
+extension FetchRequestTests {
     func testRxFetchOneRow() throws {
         try Test(testRxFetchOneRow)
             .run { try DatabaseQueue(path: $0) }
@@ -360,8 +255,6 @@ extension TypedRequestTests {
     func testRxFetchOneRow(writer: DatabaseWriter, disposeBag: DisposeBag) throws {
         let request = SQLRequest<Row>("SELECT * FROM players ORDER BY name")
         let expectedNames = [
-            "Arthur",
-            "Arthur",
             "Arthur",
             nil,
             "Craig",
@@ -385,40 +278,7 @@ extension TypedRequestTests {
     }
 }
 
-extension TypedRequestTests {
-    func testRxFetchOneRowDistinctUntilChanged() throws {
-        try Test(testRxFetchOneRowDistinctUntilChanged)
-            .run { try DatabaseQueue(path: $0) }
-            .run { try DatabasePool(path: $0) }
-    }
-    
-    func testRxFetchOneRowDistinctUntilChanged(writer: DatabaseWriter, disposeBag: DisposeBag) throws {
-        let request = SQLRequest<Row>("SELECT * FROM players ORDER BY name")
-        let expectedNames = [
-            "Arthur",
-            nil,
-            "Craig",
-            ]
-        
-        try setUpDatabase(in: writer)
-        let recorder = EventRecorder<Row?>(expectedEventCount: expectedNames.count)
-        request.rx.fetchOne(in: writer, distinctUntilChanged: true)
-            .subscribe { event in
-                // events are expected on the main thread by default
-                assertMainQueue()
-                recorder.on(event)
-            }
-            .disposed(by: disposeBag)
-        try modifyDatabase(in: writer)
-        wait(for: recorder, timeout: 1)
-        
-        for (event, name) in zip(recorder.recordedEvents, expectedNames) {
-            XCTAssertEqual(event.element!?["name"] as String?, name)
-        }
-    }
-}
-
-extension TypedRequestTests {
+extension FetchRequestTests {
     func testRxFetchOneRowIdentifiedById() throws {
         try Test(testRxFetchOneRowIdentifiedById)
             .run { try DatabaseQueue(path: $0) }
@@ -454,7 +314,7 @@ extension TypedRequestTests {
 
 // MARK: - DatabaseValue
 
-extension TypedRequestTests {
+extension FetchRequestTests {
     func testRxFetchAllDatabaseValues() throws {
         try Test(testRxFetchAllDatabaseValues)
             .run { try DatabaseQueue(path: $0) }
@@ -464,7 +324,6 @@ extension TypedRequestTests {
     func testRxFetchAllDatabaseValues(writer: DatabaseWriter, disposeBag: DisposeBag) throws {
         let request = SQLRequest<String>("SELECT name FROM players ORDER BY name")
         let expectedNames = [
-            ["Arthur", "Barbara"],
             ["Arthur", "Barbara"],
             ["Arthur", "Barbie"],
             [],
@@ -489,41 +348,7 @@ extension TypedRequestTests {
     }
 }
 
-extension TypedRequestTests {
-    func testRxFetchAllDatabaseValuesDistinctUntilChanged() throws {
-        try Test(testRxFetchAllDatabaseValuesDistinctUntilChanged)
-            .run { try DatabaseQueue(path: $0) }
-            .run { try DatabasePool(path: $0) }
-    }
-    
-    func testRxFetchAllDatabaseValuesDistinctUntilChanged(writer: DatabaseWriter, disposeBag: DisposeBag) throws {
-        let request = SQLRequest<String>("SELECT name FROM players ORDER BY name")
-        let expectedNames = [
-            ["Arthur", "Barbara"],
-            ["Arthur", "Barbie"],
-            [],
-            ["Craig", "David", "Elena"],
-            ]
-        
-        try setUpDatabase(in: writer)
-        let recorder = EventRecorder<[String]>(expectedEventCount: expectedNames.count)
-        request.rx.fetchAll(in: writer, distinctUntilChanged: true)
-            .subscribe { event in
-                // events are expected on the main thread by default
-                assertMainQueue()
-                recorder.on(event)
-            }
-            .disposed(by: disposeBag)
-        try modifyDatabase(in: writer)
-        wait(for: recorder, timeout: 1)
-        
-        for (event, names) in zip(recorder.recordedEvents, expectedNames) {
-            XCTAssertEqual(event.element!, names)
-        }
-    }
-}
-
-extension TypedRequestTests {
+extension FetchRequestTests {
     func testRxFetchOneDatabaseValue() throws {
         try Test(testRxFetchOneDatabaseValue)
             .run { try DatabaseQueue(path: $0) }
@@ -533,8 +358,6 @@ extension TypedRequestTests {
     func testRxFetchOneDatabaseValue(writer: DatabaseWriter, disposeBag: DisposeBag) throws {
         let request = SQLRequest<String>("SELECT name FROM players ORDER BY name")
         let expectedNames = [
-            "Arthur",
-            "Arthur",
             "Arthur",
             nil,
             "Craig",
@@ -558,42 +381,9 @@ extension TypedRequestTests {
     }
 }
 
-extension TypedRequestTests {
-    func testRxFetchOneDatabaseValueDistinctUntilChanged() throws {
-        try Test(testRxFetchOneDatabaseValueDistinctUntilChanged)
-            .run { try DatabaseQueue(path: $0) }
-            .run { try DatabasePool(path: $0) }
-    }
-    
-    func testRxFetchOneDatabaseValueDistinctUntilChanged(writer: DatabaseWriter, disposeBag: DisposeBag) throws {
-        let request = SQLRequest<String>("SELECT name FROM players ORDER BY name")
-        let expectedNames = [
-            "Arthur",
-            nil,
-            "Craig",
-            ]
-        
-        try setUpDatabase(in: writer)
-        let recorder = EventRecorder<String?>(expectedEventCount: expectedNames.count)
-        request.rx.fetchOne(in: writer, distinctUntilChanged: true)
-            .subscribe { event in
-                // events are expected on the main thread by default
-                assertMainQueue()
-                recorder.on(event)
-            }
-            .disposed(by: disposeBag)
-        try modifyDatabase(in: writer)
-        wait(for: recorder, timeout: 1)
-        
-        for (event, name) in zip(recorder.recordedEvents, expectedNames) {
-            XCTAssertEqual(event.element!, name)
-        }
-    }
-}
-
 // MARK: - Optional DatabaseValue
 
-extension TypedRequestTests {
+extension FetchRequestTests {
     func testRxFetchAllOptionalDatabaseValues() throws {
         try Test(testRxFetchAllOptionalDatabaseValues)
             .run { try DatabaseQueue(path: $0) }
@@ -603,8 +393,6 @@ extension TypedRequestTests {
     func testRxFetchAllOptionalDatabaseValues(writer: DatabaseWriter, disposeBag: DisposeBag) throws {
         let request = SQLRequest<String?>("SELECT email FROM players ORDER BY name")
         let expectedNames = [
-            ["arthur@example.com", nil],
-            ["arthur@example.com", nil],
             ["arthur@example.com", nil],
             [],
             [nil, "david@example.com", "elena@example.com"],
@@ -632,44 +420,7 @@ extension TypedRequestTests {
     }
 }
 
-extension TypedRequestTests {
-    func testRxFetchAllOptionalDatabaseValuesDistinctUntilChanged() throws {
-        try Test(testRxFetchAllOptionalDatabaseValuesDistinctUntilChanged)
-            .run { try DatabaseQueue(path: $0) }
-            .run { try DatabasePool(path: $0) }
-    }
-    
-    func testRxFetchAllOptionalDatabaseValuesDistinctUntilChanged(writer: DatabaseWriter, disposeBag: DisposeBag) throws {
-        let request = SQLRequest<String?>("SELECT email FROM players ORDER BY name")
-        let expectedNames = [
-            ["arthur@example.com", nil],
-            [],
-            [nil, "david@example.com", "elena@example.com"],
-            ]
-        
-        try setUpDatabase(in: writer)
-        let recorder = EventRecorder<[String?]>(expectedEventCount: expectedNames.count)
-        request.rx.fetchAll(in: writer, distinctUntilChanged: true)
-            .subscribe { event in
-                // events are expected on the main thread by default
-                assertMainQueue()
-                recorder.on(event)
-            }
-            .disposed(by: disposeBag)
-        try modifyDatabase(in: writer)
-        wait(for: recorder, timeout: 1)
-        
-        for (event, names) in zip(recorder.recordedEvents, expectedNames) {
-            let fetchedNames = event.element!
-            XCTAssertEqual(fetchedNames.count, names.count)
-            for (fetchedName, name) in zip(fetchedNames, names) {
-                XCTAssertEqual(fetchedName, name)
-            }
-        }
-    }
-}
-
-extension TypedRequestTests {
+extension FetchRequestTests {
     func testPrimaryKeyDiffScanner() throws {
         try Test(testPrimaryKeyDiffScanner)
             .run { try DatabaseQueue(path: $0) }
@@ -684,10 +435,6 @@ extension TypedRequestTests {
                     Player(id: 1, name: "Arthur", email: "arthur@example.com"),
                     Player(id: 2, name: "Barbara", email: nil),
                     ],
-                updated: [],
-                deleted: []),
-            PrimaryKeyDiff(
-                inserted: [],
                 updated: [],
                 deleted: []),
             PrimaryKeyDiff(